chore: expand ruff lint rule coverage
Adopt additional ruff rule groups beyond the existing E/W/F/I/B/C4/UP/BLE:
FA, ISC, ICN, RET, SIM, TID, PTH, TD, NPY, RUF, PERF, PT, ASYNC, PGH, PIE,
LOG, G. Apply autofixes and resolve all remaining violations by hand. Bump
the ruff pre-commit hook to v0.15.9 to match the project's pinned ruff so it
recognizes the newer rule selectors.

Scoped ignores (with rationale in pyproject.toml):
- ISC001 conflicts with the formatter
- SIM105 (contextlib.suppress) is debatable
- TD002/TD003 (TODO author/link) are not team conventions
- RUF001/002/003 ambiguous-unicode are intentional in user-facing text
- ASYNC240 advises trio/anyio Path; this project is asyncio-only

Per-file (tests/**): ASYNC109/110 (test helpers/polling), PERF (test clarity
over micro-perf), SIM117 (nested mock.patch is a readable idiom). All three
stay active for production code.

Deferred to follow-up PRs (documented, not done here): complexity guardrails
(C90 + PLR09xx), DTZ datetime-tz awareness, TC TYPE_CHECKING blocks (runtime
risk with typer/pydantic), S/bandit, and non-ruff duplicate-code tooling.

+++ transcriber/batch_processor.py
@@ -108,7 +108,7 @@ class ChecklistView:
 
 
 def _format_duration(seconds: float) -> str:
-    total = int(round(max(seconds, 0.0)))
+    total = round(max(seconds, 0.0))
     minutes, secs = divmod(total, 60)
     return f"{minutes:02d}:{secs:02d}"
